requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2

tensorflow==2.15.0
pillow==11.1.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.1
scipy==1.10.1
pandas==1.5.3
seaborn==0.12.2
scikit-learn==1.2.2

torch==2.1.0
torchvision==0.16.0
